# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Agentic Gateway API
  version: 0.3.0
  description: |
    Agentic Gateway by Dojo Genesis - A production-grade AI agent gateway with MCP integration,
    disposition-driven agent configuration, and comprehensive observability.

    ## Features
    - OpenAI-compatible chat completions API
    - MCP (Model Context Protocol) server integration
    - Agent disposition system (ADA contract)
    - DAG-based orchestration
    - Memory management with compression
    - OTEL trace export

  contact:
    name: Dojo Genesis
    url: https://github.com/DojoGenesis/gateway
servers:
  - url: http://localhost:8081
    description: Development server
  - url: https://gateway.example.com
    description: Production server
...
paths:
...
  /v1/gateway/agents/{id}:
    get:
      tags: [Gateway]
      summary: Get agent status
      description: Retrieve current agent configuration and status
      operationId: getAgent
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Agent details
          content:
            application/json:
              schema:
                type: object
                properties:
                  agent_id:
                    type: string
                  config:
                    $ref: '#/components/schemas/AgentConfig'
                  status:
                    type: string
                    enum: [active, paused, stopped]
        '404':
          $ref: '#/components/responses/NotFound'
...
  /v1/gateway/traces/{id}:
    get:
      tags: [Gateway]
      summary: Get trace details
      description: Retrieve OTEL trace information (if OTEL enabled)
      operationId: getTrace
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Trace details
          content:
            application/json:
              schema:
                type: object
                properties:
                  trace_id:
                    type: string
                  spans:
                    type: array
                    items:
                      type: object

  # ==================== Admin Routes ====================
...
components:
  schemas:
    AgentConfig:
      type: object
      description: Agent disposition configuration
      properties:
        pacing:
          type: string
          enum: [deliberate, measured, responsive, rapid]
          description: Execution pacing preference
        depth:
          type: string
          enum: [surface, functional, thorough, exhaustive]
          description: Analysis depth preference
        tone:
          type: string
          enum: [formal, professional, conversational, casual]
        initiative:
          type: string
          enum: [reactive, responsive, proactive, autonomous]
...
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
...
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: Authorization
      description: API key authentication
security:
  - ApiKeyAuth: []